+++ src/modules/helpers/editor/vscode.ts
@@ -1,3 +1,5 @@
+import type { BaseEditorHelperOptions } from 'typings/editor.ts'
+
 import { createEditorFileConfig } from 'modules/helpers/editor/main.ts'
 import { getConfigDir } from 'modules/helpers/paths.ts'
 
@@ -5,12 +7,15 @@ import { getConfigDir } from 'modules/helpers/paths.ts'
  * Creates a `VSCode` configuration file (`settings.json`) for the current project.
  * This function generates a configuration file specifically tailored for use with VSCode.
  *
+ * @param options The editor helper options
+ *   - `baseRoot`: The base root directory where the folder should be created. Defaults to root.
+ *
  * @category helpers
  */
-export function createVSCodeConfig(): Promise<boolean> {
+export function createVSCodeConfig(options?: BaseEditorHelperOptions): Promise<boolean> {
   const config = getConfigDir()?.split('/').pop()
   return createEditorFileConfig(
-    { type: 'vscode' },
+    { type: 'vscode', ...options },
     (content) => content.replace('$DENO_CONFIG', config || 'deno.json'),
   )
 }