+++ src/modules/helpers/github/workflows/main.ts
@@ -5,6 +5,7 @@ import { getRootDir } from 'modules/helpers/paths.ts'
 import { capitalize } from 'utils/strings.ts'
 import logger from 'modules/logger/mod.ts'
 import constants from 'utils/constants.ts'
+import { join } from '@std/path'
 
 /** Base function to create a YAML workflow file */
 export async function createWorkflow(
@@ -14,26 +15,28 @@ export async function createWorkflow(
   replaceContentCallback: (content: string) => string = (content) => content,
 ) {
   if (!options.filename) {
-    logger.warn('No workflow YAML file found for this project, unable to create', 'noSave')
+    logger.warn('No workflow YAML file found for this project, skipping creation.', 'noSave')
     return false
   }
 
-  const root = getRootDir()
-  const { baseFolder = `${root}/${constants.GITHUB_WORKFLOW_FOLDER}`, filename: yml } = options
+  const { baseFolder = constants.GITHUB_WORKFLOW_FOLDER, baseRoot = getRootDir(), filename: yml } =
+    options
+
   const mainYamls = capitalize(yml)
+  const dir = join(baseRoot, baseFolder)
 
   try {
     // Create content for the pre-commit hook
     const hookContent = await readFileFromCurrentUrl(import.meta.url, `./yamls/${yml}.base.yml`)
 
     // Create the .github/workflow directory if it doesn't exist
-    await Deno.mkdir(baseFolder, { recursive: true })
+    await Deno.mkdir(dir, { recursive: true })
 
     // file dir
-    const baseFileDir = `${baseFolder}/${yml}.yml`
+    const baseFileDir = `${dir}/${yml}.yml`
 
     if (fileExists(baseFileDir)) {
-      logger.warn(`'${mainYamls}' YAML already exists, unable to create or update`, 'noSave')
+      logger.warn(`'${mainYamls}' YAML already exists, skipping creation.`, 'noSave')
       return false
     }
 
@@ -44,7 +47,7 @@ export async function createWorkflow(
 
     return true
   } catch (e) {
-    logger.error(`'${mainYamls}' YAML creation error in '${baseFolder}'`, e, 'noSave')
+    logger.error(`'${mainYamls}' YAML creation error in '${dir}'`, e, 'noSave')
 
     return false
   }